beatles(LearnEDA)	R Documentation

Lengths of songs from Beatles' albums.

Description

Length of songs from a selection of albums from the popular pop and rock group The Beatles.

Usage


beatles

Format

A data frame with 113 observations on the following 2 variables.

time

time in seconds of the song

album

name of Beatles' album

Source

Collected by the author from the album descriptions

[Package LearnEDA version 1.0 Index]