boston.marathon(LearnEDA)	R Documentation

Boston Marathon completion times of women of different ages.

Description

Completion times for a sample of women at the 2001 Boston Marathon.

Usage


boston.marathon

Format

A data frame with 108 observations on the following 2 variables.

time

completion time in minutes

age

age of the participant

Source

www.bostonmarathon.org website

[Package LearnEDA version 1.0 Index]