college.ratings(LearnEDA)	R Documentation

Ratings of National Universities in the U.S.

Description

Ratings for National Universities based on a 2001 survey.

Usage


college.ratings

Format

A data frame with 249 observations on the following 12 variables.

School

name of university

Tier

tier grouping of the univesity

Type

private or public institution

Reputation

measure of academic reputation

F.retention

average freshmen retention rate

a.grad.rate

percentage of freshmen who graduated within a six-year period

Under.20

percentage of undergraduate classes with fewer than 20 students

Over.50

percentage of undergraduate classes with 50 students or more enrolled

Full.time

percentage of the total number of faculty employed on a full-time basis

Top.10

proportion of students enrolled who graduated in the top 10 percent of their high school class

Acc.rate

ratio of the number of students admitted to the number of applicants

Alum.giving

average percent of undergraduate alumni of record who donated money to the college or university

Source

U.S. News and World Report

[Package LearnEDA version 1.0 Index]