half.slope.ratio(LearnEDA)	R Documentation

Half slope ratio

Description

Computes the half slope ratio for three summary points and power transformations on the two variables.

Usage


half.slope.ratio(sx,sy,px,py)

Arguments

sx 	 numeric vector of summary values of the x variable

sy 	 numeric vector of summary values of the y variable

px 	 power of the transformation on the x variable

py 	 power of the transformation on the y variable

Value

Value of the half slope ratio on the transformed summary points

Author(s)

Jim Albert

Examples


sx=c(10,30,50)
sy=c(5,8,20)
half.slope.ratio(sx,sy,1,1)
half.slope.ratio(sx,sy,-.5,-.5)

[Package LearnEDA version 1.0 Index]