homeruns.2000(LearnEDA)	R Documentation

Team home run numbers for different seasons.

Description

Total number of home runs hit all teams in the major league baseball league for the years 1900,1910,..., 2000.

Usage


homeruns.2000

Format

A data frame with 210 observations on the following 2 variables.

YEARS

season

HOMERUNS

count of team home runs

Source

Sean Lahman's database at www.baseball1.com

[Package LearnEDA version 1.0 Index]