homeruns.61(LearnEDA)	R Documentation

Home run counts in 1961

Description

The number of the number of home runs hit by all major league players in the year 1961.

Usage


homeruns.61

Format

A data frame with 698 observations on the following 1 variables.

HR

count of home runs

Source

Sean Lahman's database at www.baseball1.com

[Package LearnEDA version 1.0 Index]