island.areas(LearnEDA)	R Documentation

Areas of islands from different continents.

Description

Total number of home runs hit all teams in the major league baseball league for the years 1900,1910,..., 2000.

Usage


islands.areas

Format

A data frame with 59 observations on the following 3 variables.

Ocean

ocean where the island lies

Name

name of the island

Area

area of the island

Source

World Almanac and Book of Facts

[Package LearnEDA version 1.0 Index]