lval(LearnEDA)	R Documentation

Letter values

Description

Computes the letter values for a batch of numbers.

Usage


lval(x,na.rm=TRUE)

Arguments

x 	 numeric vector of values

na.rm 	 logical value indicating whether NA values should be stripped before the computation proceeds

Value

dataframe with components depth, lo, hi, mids, and spreads

Author(s)

Jim Albert

Examples


lval(rnorm(100))

[Package LearnEDA version 1.0 Index]